+++ CnCTDRAMapEditor/Properties/Settings.Designer.cs
@@ -197,6 +197,15 @@ namespace MobiusEditor.Properties {
         [global::System.Configuration.ApplicationScopedSettingAttribute()]
         [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
         [global::System.Configuration.DefaultSettingValueAttribute("True")]
+        public bool Ignore106Scripting {
+            get {
+                return ((bool)(this["Ignore106Scripting"]));
+            }
+        }
+        
+        [global::System.Configuration.ApplicationScopedSettingAttribute()]
+        [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
+        [global::System.Configuration.DefaultSettingValueAttribute("True")]
         public bool NoMetaFilesForSinglePlay {
             get {
                 return ((bool)(this["NoMetaFilesForSinglePlay"]));