+++ CnCTDRAMapEditor/Tools/Dialogs/OverlayToolDialog.cs
@@ -19,7 +19,7 @@ namespace MobiusEditor.Tools.Dialogs
         protected override void InitializeInternal(MapPanel mapPanel, MapLayerFlag activeLayers, ToolStripStatusLabel toolStatusLabel, ToolTip mouseToolTip, IGamePlugin plugin, UndoRedoList<UndoRedoEventArgs> undoRedoList)
         {
             GenericTypeListBox.Types = plugin.Map.OverlayTypes.
-                Where(t => t.IsPlaceable && (!Globals.FilterTheaterObjects || t.Theaters == null || t.Theaters.Contains(plugin.Map.Theater))).
+                Where(t => t.IsOverlay && (!Globals.FilterTheaterObjects || t.Theaters == null || t.Theaters.Contains(plugin.Map.Theater))).
                 OrderBy(t => t.ID);
 
             Tool = new OverlaysTool(mapPanel, activeLayers, toolStatusLabel,